# Improve /info performance by adding a index to highlights