--- a/db/migrate/20200622170535_add_user_id_source_id_index_to_highlights.rb
+++ b/db/migrate/20200622170535_add_user_id_source_id_index_to_highlights.rb
@@ -1,0 +1,5 @@
+class AddUserIdSourceIdIndexToHighlights < ActiveRecord::Migration[5.2]
+  def change
+    add_index :highlights, [:user_id, :source_id]
+  end
+end