--- a/public/profiling.html
+++ b/public/profiling.html
@@ -1,0 +1,16 @@
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Page title</title>
+    <link rel="stylesheet" href="css/style.css">
+</head>
+<body>
+
+To use mini-profiler on a json page, run the json query first.  Then, run this page.
+This works because mini-profile caches the json query in memory, but needs a html/body page
+in order to show the results.
+
+</body>
+</html>